requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==4.2.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dependency-injector==4.42.0b1
distro==1.9.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
groq==0.9.0
grpcio==1.65.2
grpcio-tools==1.65.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
MarkupSafe==2.1.5
matrix-nio==0.24.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.18.1
packaging==24.1
pathspec==0.12.1
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
priority==2.0.0
protobuf==5.27.3
pycparser==2.22
pycryptodome==3.20.0
pycurl==7.45.3
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-olm==3.2.16
python-socks==2.5.0
PyYAML==6.0.1
qdrant-client==1.10.1
Quart==0.19.6
rake-nltk==1.0.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
segtok==1.5.11
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
unpaddedbase64==2.1.0
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
wsproto==1.2.0
yake==0.4.8
yarl==1.9.4
